Moved IVP examples into separate file
